split tests/ieee.rs into 2 parts
